rename auth middleware
